pyproject.toml
[project]
name = "mojiokoshi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "numpy<2.0",
    "setuptools>=81.0.0",
    "soundfile>=0.13.1",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "wheel>=0.46.3",
    "openai-whisper",
    "numba",
    "tkinterdnd2",
    "pyannote-audio>=3.4.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.18.0",
]
